dbh_estimation_algorithm_FastAPI/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
flatbuffers==23.5.26
fonttools==4.45.1
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.2
pandas==2.0.3
patsy==0.5.3
Pillow==10.1.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
statsmodels==0.14.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0rc0
tensorflow-estimator==2.13.0
termcolor==2.3.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
